requirements.txt
Django==4.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7